// src/ts/Enka.Network_icon2text/types/artifactOptionKey.ts
export type artifactMainOptionKey
    = "HP"
    | "ATTACK"
    | "HP_PERCENT"
    | "ATTACK_PERCENT"
    | "DEFENSE_PERCENT"
    | "CRITICAL"
    | "CRITICAL_HURT"
    | "CHARGE_EFFICIENCY"
    | "HEAL_ADD"
    | "ELEMENT_MASTERY"
    | "PHYSICAL_ADD_HURT"
    | "FIRE_ADD_HURT"
    | "ELEC_ADD_HURT"
    | "WATER_ADD_HURT"
    | "WIND_ADD_HURT"
    | "ICE_ADD_HURT"
    | "ROCK_ADD_HURT"
    | "GRASS_ADD_HURT"
    | "UNKNOWN";

export type artifactSubOptionKey
    = "HP"
    | "ATTACK"
    | "DEFENSE"
    | "HP_PERCENT"
    | "ATTACK_PERCENT"
    | "DEFENSE_PERCENT"
    | "CRITICAL"
    | "CRITICAL_HURT"
    | "CHARGE_EFFICIENCY"
    | "ELEMENT_MASTERY"
    | "UNKNOWN";